chore: rename `tron-client.ts` to `client.ts`
